Restructure project to be module based
