+++ tools/machine-learning/multi-task-yolo/pyproject.toml
@@ -13,3 +13,7 @@ dependencies = [
   "torch>=2.10.0",
   "ultralytics>=8.4.31",
 ]
+
+[build-system]
+requires = ["setuptools>=68", "wheel"]
+build-backend = "setuptools.build_meta"
