+++ tools/machine-learning/multi-task-yolo/src/model/multi_task_yolo.py
