+++ tools/machine-learning/multi-task-yolo/src/validation/assets/nao_coco_k1_data.yaml
