+++ tools/machine-learning/multi-task-yolo/src/validation/assets/yolo26m-tuned.pt
